Remove auth component
